Add beacon detection
We only want to fetch once from every beacon each sleep cycle.
Therefore, there has to be a check on whether or not the beacon has been
detected this sleep cycle.

This commit introduces an array that contains addresses with the type
`mac_address_t`.

Note that the array or the `discovered_beacon_count` doesn't reset. This
should be done in a coming commit.